// src/app/Model/LoginModel.ts
import {PristineNavigationItem} from "../../@pristinebase/types";

export class LoginModel {
    name: string;
    email: string;
    roleId: string;
    condition: string;
    message: string;
    menu: Array<PristineNavigationItem>;
    shiftID: string;
    workType:string;
    phoneNo: string;
    email_ids: string;
    locationCode: string;
}